Move to spago@next
The sources now build with spago@next.  The tests do not run because if I
set up the spago.yaml properly, spago complains of a naming conflict.  I
suspect this is because the docs examples all contain a main method.

Currently I have it that the spago.yaml in the docs/src diectory (which
was docs/Examples) uses a different resource name from that used in the
workspace as an extra path.  This allows building but, of course, not
testing.